Changed the message when calling the Start command Air quality has been added to the current weather description Added ENUM UsEpaIndex.java The parameter &aqi=yes has been added to the HTTP request

+++ src/main/java/com/app/weatherGPT/dto/api/weather/AirQuality.java
@@ -13,12 +13,12 @@ import java.io.Serializable;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AirQuality implements Serializable {
 
-    private double co;
-    private double no2;
-    private double o3;
-    private double so2;
-    private double pm2_5;
-    private double pm10;
-    private double us_epa_index;
-    private double gb_defra_index;
+    private float co;
+    private float no2;
+    private float o3;
+    private float so2;
+    private float pm2_5;
+    private float pm10;
+    private int us_epa_index;
+    private int gb_defra_index;
 }
